improve type cleanness
this silences the following compiler warning on newer gccs

sockfile.c: In function ‘Ns_ResetFileVec’:
sockfile.c:126:32: warning: conversion to ‘long unsigned int’ from ‘off_t’ {aka ‘long int’} may change the sign of the result [-Wsign-conversion]
  126 |                 bufs[i].offset += sent;
      |                                ^~
sockfile.c:126:35: warning: conversion to ‘off_t’ {aka ‘long int’} from ‘long unsigned int’ may change the sign of the result [-Wsign-conversion]
  126 |                 bufs[i].offset += sent;

+++ nsd/sockfile.c
@@ -112,10 +112,11 @@ Ns_SetFileVec(Ns_FileVec *bufs, int i,  int fd, const void *data,
 int
 Ns_ResetFileVec(Ns_FileVec *bufs, int nbufs, size_t sent)
 {
-    int          i;
+    int i;
 
     for (i = 0; i < nbufs && sent > 0u; i++) {
         size_t length = bufs[i].length;
+
         if (length > 0u) {
             if (sent >= length) {
                 sent -= length;
@@ -123,7 +124,7 @@ Ns_ResetFileVec(Ns_FileVec *bufs, int nbufs, size_t sent)
                 bufs[i].offset = 0;
                 bufs[i].length = 0;
             } else {
-                bufs[i].offset += sent;
+                bufs[i].offset += (off_t)sent;
                 bufs[i].length -= sent;
                 break;
             }
